Created trackWidget.js
