+++ index.html
@@ -149,8 +149,7 @@
                                     Local File ...
                                 </div>
                                 <div>
-                                    <input id="igv-app-dropdown-local-track-file-input" type="file" name="file" multiple
-                                           style="display: none;">
+                                    <input id="igv-app-dropdown-local-track-file-input" type="file" name="file" multiple style="display: none;">
                                 </div>
                             </div>
                         </label>
@@ -180,11 +179,12 @@
                         </div>
 
                         <!-- URL -->
-                        <button class="dropdown-item" type="button" data-toggle="modal"
-                                data-target="#igv-app-track-from-url-modal">URL ...
+                        <button class="dropdown-item" type="button" data-toggle="modal" data-target="#igv-app-track-from-url-modal">
+                            URL ...
                         </button>
 
-                        <div id="igv-app-annotations-section" class="dropdown-divider"></div>
+                        <div id="igv-app-annotations-section" class="dropdown-divider">
+                        </div>
 
                     </div>
                 </div>
@@ -324,38 +324,10 @@
         </div>
     </div>
 
-    <!-- Multiple File Load Modal -->
-    <div id="igv-app-multiple-file-load-modal" class="modal" tabindex="-1" role="dialog"
-         aria-labelledby="igv-app-multiple-file-load-modal-label" aria-hidden="true">
-
-        <div class="modal-dialog modal-lg" role="document">
-
-            <div class="modal-content">
-
-                <div class="modal-header">
-                    <div id="igv-app-multiple-file-load-modal-label" class="modal-title">
-                    </div>
-
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-
-                </div>
-
-                <div id="igv-app-multiple-file-load-modal-body" class="modal-body">
-                </div>
-
-            </div>
-
-        </div>
-
-    </div>
-
     <!-- Generic Track Select Modal -->
-    <div id="igv-app-generic-track-select-modal" class="modal fade" tabindex="-1" role="dialog"
-         aria-labelledby="igv-app-generic-track-select-modal-label" aria-hidden="true">
+    <div id="igv-app-generic-track-select-modal" class="modal fade">
 
-        <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-dialog modal-lg">
 
             <div class="modal-content">
 
@@ -363,7 +335,7 @@
                     <div id="igv-app-generic-track-select-modal-label" class="modal-title">
                     </div>
 
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <button type="button" class="close" data-dismiss="modal">
                         <span aria-hidden="true">&times;</span>
                     </button>
 
@@ -385,17 +357,16 @@
     </div>
 
     <!-- Genome from URL -->
-    <div id="igv-app-genome-from-url-modal" class="modal fade" tabindex="-1" role="dialog"
-         aria-labelledby="igv-app-genome-from-url-modal-label" aria-hidden="true">
+    <div id="igv-app-genome-from-url-modal" class="modal fade">
 
-        <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-dialog modal-lg">
 
             <div class="modal-content">
 
                 <div class="modal-header">
                     <div id="igv-app-genome-from-url-modal-label" class="modal-title">Genome URL</div>
 
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <button type="button" class="close" data-dismiss="modal">
                         <span aria-hidden="true">&times;</span>
                     </button>
 
@@ -416,17 +387,16 @@
     </div>
 
     <!-- Track from URL -->
-    <div id="igv-app-track-from-url-modal" class="modal fade" tabindex="-1" role="dialog"
-         aria-labelledby="igv-app-track-from-url-modal-label" aria-hidden="true">
+    <div id="igv-app-track-from-url-modal" class="modal fade">
 
-        <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-dialog modal-lg">
 
             <div class="modal-content">
 
                 <div class="modal-header">
                     <div id="igv-app-track-from-url-modal-label" class="modal-title">Track URL</div>
 
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <button type="button" class="close" data-dismiss="modal">
                         <span aria-hidden="true">&times;</span>
                     </button>
