+++ js/app.js
@@ -22,24 +22,20 @@
  */
 
 import igv from '../node_modules/igv/dist/igv.esm.js';
-import { Alert, EventBus, Utils, GoogleFilePicker, MultipleTrackFileLoad, FileLoadManager, FileLoadWidget, SessionController, SessionFileLoad } from '../node_modules/igv-widgets/dist/igv-widgets.js';
-import { ModalTable } from '../node_modules/data-modal/js/index.js';
+import { Alert, EventBus, GoogleFilePicker, SessionController, SessionFileLoad } from '../node_modules/igv-widgets/dist/igv-widgets.js';
 import Globals from "./globals.js"
-import { updateTrackMenus } from './trackMenu.js';
 import { creatGenomeWidgets, initializeGenomeWidgets, genomeWidgetConfigurator } from './genomeWidgets.js';
 import { shareWidgetConfigurator, createShareWidgets } from './shareWidgets.js';
 import { sessionURL } from './shareHelper.js';
 import { createSVGWidget } from './svgWidget.js';
+import {createTrackWidgets} from "./trackWidgets.js";
 
 $(document).ready(async () => main($('#igv-app-container'), igvwebConfig));
 
-let eventBus = new EventBus();
-
-let fileLoadWidget;
-let multipleTrackFileLoad;
-let encodeModalTable;
-let trackLoadController;
 let sessionController;
+
+let eventBus = new EventBus();
+let trackLoadController;
 let googleEnabled = false;
 
 let main = async ($container, config) => {
@@ -95,7 +91,7 @@ let initializationHelper = async (browser, $container, options) => {
     creatGenomeWidgets(genomeWidgetConfigurator())
     await initializeGenomeWidgets(browser, options.genomes, $('#igv-app-genome-dropdown-menu'))
 
-    await createTrackLoadGUI(browser, googleEnabled, igv.xhr, igv.google, options);
+    await createTrackWidgets(browser, googleEnabled, igv.xhr, igv.google, options);
 
     createSessionSaveLoadGUI(browser);
 
@@ -104,75 +100,6 @@ let initializationHelper = async (browser, $container, options) => {
     createShareWidgets(shareWidgetConfigurator(browser, $container, options));
 
     createAppBookmarkHandler($('#igv-app-bookmark-button'));
-
-}
-
-const createTrackLoadGUI = async (browser, googleEnabled, igvxhr, google, { trackRegistryFile }) => {
-
-    const $urlModal = $('#igv-app-track-from-url-modal')
-
-    let fileLoadWidgetConfig =
-        {
-            widgetParent: $urlModal.find('.modal-body').get(0),
-            dataTitle: 'Track',
-            indexTitle: 'Track Index',
-            mode: 'url',
-            fileLoadManager: new FileLoadManager(),
-            dataOnly: false,
-            doURL: true
-        };
-
-    fileLoadWidget = new FileLoadWidget(fileLoadWidgetConfig)
-
-    Utils.configureModal(fileLoadWidget, $urlModal.get(0), async fileLoadWidget => {
-        const paths = fileLoadWidget.retrievePaths();
-        await multipleTrackFileLoad.loadPaths( paths );
-        return true;
-    });
-
-    let $igv_app_dropdown_google_drive_track_file_button = $('#igv-app-dropdown-google-drive-track-file-button');
-    if (!googleEnabled) {
-        $igv_app_dropdown_google_drive_track_file_button.parent().hide();
-    }
-
-    const $googleDriveButton = googleEnabled ? $igv_app_dropdown_google_drive_track_file_button : undefined;
-
-    const multipleTrackFileLoadConfig =
-        {
-            $localFileInput: $('#igv-app-dropdown-local-track-file-input'),
-            $dropboxButton: $('#igv-app-dropdown-dropbox-track-file-button'),
-            $googleDriveButton,
-            fileLoadHandler: async configurations => await browser.loadTrackList(configurations),
-            multipleFileSelection: true,
-            igvxhr,
-            google
-        };
-
-    multipleTrackFileLoad = new MultipleTrackFileLoad(multipleTrackFileLoadConfig)
-
-    const encodeModalTableConfig =
-        {
-            id: "igv-app-encode-modal",
-            title: "ENCODE",
-            selectionStyle: 'multi',
-            pageLength: 100,
-            selectHandler: async configurations => await browser.loadTrackList( configurations )
-        };
-
-    encodeModalTable = new ModalTable(encodeModalTableConfig)
-
-    await updateTrackMenus(browser.genome.id, encodeModalTable, trackRegistryFile, $('#igv-app-track-dropdown-menu'), $('#igv-app-generic-track-select-modal'), configuration => browser.loadTrack(configuration));
-
-    const genomeChangeListener = {
-
-        receiveEvent: async ({ data }) => {
-
-            const { genomeID } = data;
-            await updateTrackMenus(genomeID, encodeModalTable, trackRegistryFile, $('#igv-app-track-dropdown-menu'), $('#igv-app-generic-track-select-modal'), configuration => browser.loadTrack(configuration));
-        }
-    }
-
-    eventBus.subscribe('DidChangeGenome', genomeChangeListener);
 
 }
 
@@ -225,4 +152,4 @@ const createAppBookmarkHandler = $bookmark_button => {
 
 }
 
-export { main, eventBus, googleEnabled, encodeModalTable, trackLoadController }
+export { main, eventBus, googleEnabled, trackLoadController }
